corrected structs

+++ models/Message.go
@@ -2,10 +2,10 @@ package models
 
 type Message struct {
 	MessageId int
-	Body string
-	CreatedAt int
-	SenderId int
-	TargetId int
-	Type string
+	Body      string
+	CreatedAt int64
+	SenderId  int
+	TargetId  int
+	Type      string
 	UpdatedAt int64
 }
